project MVP 모든 값 입력 확인 기능

+++ front/src/components/project/Project.js
@@ -3,17 +3,38 @@ import { Card, Button, Form, Row, Col } from "react-bootstrap";
 import ProjectEditForm from "./ProjectEditForm"
 import * as Api from "../../api";
 
-function Project({ project, setProjects, projects }){
+function Project({ project, setProjects, projects, key }){
 
     const [isEditing, setIsEditing] = useState(false)
 
-    const deletePost = async () => {
-        await Api.delete(`project/${project._id}`);
-        setProjects(
-          projects.filter((obj) => {
-             obj._id !== project._id;
-          })
-        )
+    // 편집창 열기
+    // const openEdit = () =>{
+      
+    //   const compareId = projects.forEach((project)=>{
+    //     project._id === key
+
+    //   })
+    // }
+
+
+    const deletePost = async(e) => {
+      e.preventDefault()
+      const confirmDelete = window.confirm("정말로 삭제하시겠습니까?")
+      if (confirmDelete == true){
+      const res = await Api.delete(`project/${project._id}`);
+      
+        if (res.data.message === "It's deleted!") {
+          setProjects(
+            projects.filter((project) => {
+              project._id !== key;
+            })
+          ) 
+          alert("삭제되었습니다")
+        }    
+      }
+
+     
+       
       }
 
     return(
@@ -35,10 +56,14 @@ function Project({ project, setProjects, projects }){
                 </Col>
 
                 <Col xs={2} sm={{ span: 20 }} >
-                   <Button variant="outline-info" size="sm" onClick={()=>{
+                   <Button 
+                   variant="outline-info" size="sm" 
+                   onClick={()=>{
                        setIsEditing(true)
                    }}>편집</Button>
-                   <Button type="submit" variant="outline-info" size="sm" onClick={deletePost}>삭제</Button>
+                   <Button 
+                   type="submit" variant="outline-info" size="sm" 
+                   onClick={deletePost}>삭제</Button>
                 </Col>
 
             </Row>
